projectStructureChanges
