--- react-text-to-speech/src/App.css
@@ -1,69 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.container {
-  display: flex;
-  align-items: center;
-  justify-content: center
-}
- 
- img {
-   max-width: 100%;
-   max-height:100%;
- }
- 
- .text {
-  font-size: 25px;
-  padding-left: 20px;
- }
-
- table {
-  border: 0px solid rgb(255, 255, 255);
-  width: 600px;
-  height: 200px;
-}
-  
-th {
-  border-bottom: 1px solid black;
-}
-  
-td {
-  text-align: center;
-}
- 
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
