+++ react-text-to-speech/src/pages/Home.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import '../styles/Home.css'
+
+function Home() {
+  return (
+    <div className='home'>
+      <div className='headerContainer' >
+        <h1>Story Telling App</h1>
+      </div>
+    </div>
+  )
+}
+
+export default Home
